drfsunscrapers/requirements.txt
alabaster==0.7.12
amqp==2.5.2
aniso8601==7.0.0
appdirs==1.4.3
arrow==0.15.4
astroid==2.3.3
asyncio==3.4.3
attrs==19.3.0
awscli==1.16.310
Babel==2.8.0
backcall==0.1.0
billiard==3.6.1.0
binaryornot==0.4.4
bleach==3.1.0
bokeh==1.4.0
boto==2.49.0
boto3==1.10.46
botocore==1.13.46
celery==4.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
chartify==2.7.0
Click==7.0
colorama==0.4.1
colour==0.1.5
cookiecutter==1.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cycler==0.10.0
Cython==0.29.14
dash==1.7.0
dash-core-components==1.6.0
dash-html-components==1.0.2
dash-renderer==1.2.2
dash-table==4.5.1
dataclasses==0.7
dateutils==0.6.6
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.8
django-cors-headers==3.2.0
django-debug-toolbar==2.1
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.5
django-haystack==2.8.1
django-nine==0.2.2
django-plotly-dash==1.1.3
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-sphinx-autodoc==0.2
django-sphinxdoc==2.0.0
django-templated-mail==1.1.1
django-webpush==0.3.1
djangorestframework==3.11.0
djongo==1.2.38
djoser==2.0.3
docutils==0.15.2
dpd-components==0.1.0
elasticsearch==7.1.0
elasticsearch-dsl==7.1.0
entrypoints==0.3
falcon==2.0.0
falcon-auth==1.1.0
falcon-swagger-ui==1.1.2
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
graphene==2.1.8
graphql-core==2.2.1
graphql-relay==2.0.1
gunicorn==20.0.4
http-ece==1.1.0
httpie==1.0.3
idna==2.8
imagesize==1.2.0
importlib-metadata==1.3.0
ios-notification==0.1.3
ipykernel==5.1.3
ipython==7.11.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
jedi==0.15.2
Jinja2==2.10.3
jinja2-time==0.2.0
jmespath==0.9.4
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kiwisolver==1.1.0
kombu==4.6.7
lazy-object-proxy==1.4.3
librabbitmq==2.0.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
mock==2.0.0
mongoengine==0.18.2
more-itertools==8.0.2
msgpack==0.6.2
msgpack-python==0.5.6
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.4
openapi-codec==1.3.2
packaging==16.8
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pbr==3.1.1
pexpect==4.7.0
pickleshare==0.7.5
pika==1.1.0
Pillow==6.2.1
plotly==4.4.1
pluggy==0.13.1
pockets==0.9.1
poyo==0.5.0
pprint==0.1
prometheus-client==0.7.1
promise==2.2.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py==1.8.0
py-vapid==1.7.0
pyasn1==0.4.8
pycparser==2.19
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
pymongo==3.10.0
pyparsing==2.2.0
pyrsistent==0.15.6
pytest==5.3.1
python-dateutil==2.6.1
python-firebase==1.2
python-logstash==0.4.6
pytz==2017.3
pywebpush==1.9.4
PyYAML==5.2
pyyml==0.0.2
pyzmq==18.1.1
qtconsole==4.6.0
rabbitmq==0.2.0
redis==3.3.11
reportlab==3.5.32
requests==2.22.0
rethinkdb==2.4.4.post1
retrying==1.3.3
rsa==3.4.2
Rx==1.6.1
s3transfer==0.2.1
scikit-learn==0.22
scipy==1.3.3
seaborn==0.9.0
selenium==3.8.0
Send2Trash==1.5.0
sentry-sdk==0.13.5
simplejson==3.17.0
six==1.11.0
smart-open==1.9.0
snowballstemmer==2.0.0
Sphinx==2.3.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-napoleon==0.7
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.11
sqlparse==0.2.4
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
uritemplate==3.0.1
urllib3==1.25.7
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
whichcraft==0.6.1
widgetsnbextension==3.5.1
wrapt==1.11.2
yellowbrick==1.0.1
zipp==0.6.0
